# api-gateway/docs/swagger.yaml
definitions:
...
  dto.Author:
    description: Author details
    properties:
      biography:
        type: string
      birth_date:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
    type: object
...
  dto.CreateAuthorRequest:
    description: A request structure for creating an author
    properties:
      biography:
        example: An accomplished author.
        type: string
      birth_date:
        example: "1980-01-01"
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
    required:
    - biography
    - birth_date
    - first_name
    - last_name
    type: object
  dto.CreateAuthorResponse:
    description: A success message for author creation
    properties:
      message:
        type: string
    type: object
...
  dto.DeleteAuthorResponse:
    description: A success message for author deletion
    properties:
      message:
        type: string
    type: object
...
  dto.ErrorResponse:
    description: Error response structure
    properties:
      message:
        type: string
    type: object
  dto.GetAuthorsResponse:
    description: A list of authors
    properties:
      authors:
        items:
          $ref: '#/definitions/dto.Author'
        type: array
    type: object
...
  dto.UpdateAuthorRequest:
    description: A request structure for updating an author
    properties:
      biography:
        example: An accomplished author.
        type: string
      birth_date:
        example: "1990-01-01"
        type: string
      first_name:
        example: John Updated
        type: string
      last_name:
        example: Doe Update
        type: string
    type: object
  dto.UpdateAuthorResponse:
    description: A success message for author update
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /authors:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of authors from the gRPC service
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        required: true
        type: integer
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetAuthorsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a list of authors
      tags:
      - Authors
    post:
      consumes:
      - application/json
      description: Create a new author with the provided information
      parameters:
      - description: Author information
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAuthorRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreateAuthorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new author
      tags:
      - Authors
  /authors/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a author with the provided information
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteAuthorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a author
      tags:
      - Authors
    put:
      consumes:
      - application/json
      description: Update a author with the provided information
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: integer
      - description: Author information
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAuthorRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateAuthorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a author
      tags:
      - Authors
...
swagger: "2.0"